requirements.txt
attrs==21.2.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
PyJWT==1.7.1
pyrsistent==0.18.0
requests==2.26.0
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.1
